+++ packages/cli-e2e/entrypoints/dockerHeadless.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 export DISPLAY=:1
 Xvfb :1 -screen 0 1024x768x16 & sleep 1
 
@@ -7,6 +9,30 @@ rsync -r --exclude="node_modules" /home/notGroot/cli/* /home/notGroot/cli-copy/
 cd /home/notGroot/cli-copy
 
 npm run setup
+npm run build
+
+export UI_TEMPLATE_VERSION=0.0.0
+npm set registry http://verdaccio:4873
+yarn config set  registry http://verdaccio:4873
+yarn config set -- --mutex network
+yarn config set -- --install.silent true
+yarn config set -- --silent true
+
+npm run npm:bump:template -- -- $UI_TEMPLATE_VERSION
+
+npm run npm:publish:template
+
+google-chrome --no-first-run --remote-debugging-port=9222 --disable-dev-shm-usage --window-size=1080,720 >/dev/null 2>&1 & \
+
+node scripts/wait-for-published-packages.js
+
 cd packages/cli-e2e
-google-chrome --no-first-run --remote-debugging-port=9222 --disable-dev-shm-usage >/dev/null 2>&1 & \
-npm run-script jest
+
+if npm run-script jest
+then
+    echo "Docker!" | sudo -S rsync -r /home/notGroot/cli-copy/packages/cli-e2e/screenshots/* /home/notGroot/cli/packages/cli-e2e/screenshots
+    exit 0
+else
+    echo "Docker!" | sudo -S rsync -r /home/notGroot/cli-copy/packages/cli-e2e/screenshots/* /home/notGroot/cli/packages/cli-e2e/screenshots
+    exit 1
+fi