+++ packages/cli/src/commands/ui/create/angular.ts
@@ -1,17 +1,30 @@
-import AuthenticationRequired from '../../../lib/decorators/authenticationRequired';
 import {Command, flags} from '@oclif/command';
 import {platformUrl} from '../../../lib/platform/environment';
-import {Storage} from '../../../lib/oauth/storage';
 import {Config} from '../../../lib/config/config';
 import {spawnProcess} from '../../../lib/utils/process';
-import {buildAnalyticsFailureHook} from '../../../hooks/analytics/analytics';
+import {
+  buildAnalyticsFailureHook,
+  buildAnalyticsSuccessHook,
+} from '../../../hooks/analytics/analytics';
 import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';
+import {getPackageVersion} from '../../../lib/utils/misc';
+import {
+  Preconditions,
+  IsAuthenticated,
+} from '../../../lib/decorators/preconditions/';
 
 export default class Angular extends Command {
+  static templateName = '@coveo/angular';
+
   static description =
-    'Create a Coveo Headless-powered search page with the Angular web framework. See https://docs.coveo.com/en/headless and https://angular.io/.';
+    'Create a Coveo Headless-powered search page with the Angular web framework. See https://docs.coveo.com/headless and https://angular.io/.';
 
   static flags = {
+    version: flags.string({
+      char: 'v',
+      description: `Version of ${Angular.templateName} to use.`,
+      default: getPackageVersion(Angular.templateName),
+    }),
     defaults: flags.boolean({
       char: 'd',
       description:
@@ -23,11 +36,16 @@ export default class Angular extends Command {
     {name: 'name', description: 'The target application name.', required: true},
   ];
 
-  @AuthenticationRequired()
+  @Preconditions(IsAuthenticated())
   async run() {
     const {args, flags} = this.parse(Angular);
     await this.createProject(args.name, flags.defaults);
     await this.addCoveoToProject(args.name, flags.defaults);
+    this.displayFeedbackAfterSuccess(args.name);
+    await this.config.runHook(
+      'analytics',
+      buildAnalyticsSuccessHook(this, flags)
+    );
   }
 
   private async createProject(name: string, defaults: boolean) {
@@ -42,16 +60,18 @@ export default class Angular extends Command {
 
   private async addCoveoToProject(applicationName: string, defaults: boolean) {
     const cfg = await this.configuration.get();
-    const storage = await this.storage.get();
     const {providerUsername} = await this.getUserInfo();
+    const {flags} = this.parse(Angular);
+    const schematicVersion =
+      flags.version || getPackageVersion(Angular.templateName);
 
     const cliArgs = [
       'add',
-      '@coveo/angular',
+      `${Angular.templateName}@${schematicVersion}`,
       '--org-id',
       cfg.organization,
       '--api-key',
-      storage.accessToken!,
+      cfg.accessToken!,
       '--platform-url',
       platformUrl({environment: cfg.environment}),
       '--user',
@@ -71,9 +91,10 @@ export default class Angular extends Command {
   }
 
   async catch(err?: Error) {
+    const {flags} = this.parse(Angular);
     await this.config.runHook(
       'analytics',
-      buildAnalyticsFailureHook(this, {}, err)
+      buildAnalyticsFailureHook(this, flags, err)
     );
     throw err;
   }
@@ -82,15 +103,23 @@ export default class Angular extends Command {
     return new Config(this.config.configDir, this.error);
   }
 
-  private get storage() {
-    return new Storage();
-  }
-
   private async getUserInfo() {
     const authenticatedClient = new AuthenticatedClient();
     const platformClient = await authenticatedClient.getClient();
     await platformClient.initialize();
 
     return await platformClient.user.get();
+  }
+
+  private displayFeedbackAfterSuccess(name: string) {
+    this.log(`
+    To get started:
+    
+    cd ${name}
+    npm run start
+    
+    See package.json for other available commands.
+    Happy hacking !
+    `);
   }
 }