+++ packages/search-token-server/middlewares/errorHandler.ts
@@ -1,6 +1,15 @@
-import type {ErrorRequestHandler} from 'express';
+import type {NextFunction, Request, Response} from 'express';
 
-export const errorHandler: ErrorRequestHandler = (err, _req, res, _next) => {
+interface HTTPError extends Error {
+  statusCode: number;
+}
+
+export function errorHandler(
+  err: HTTPError,
+  _req: Request,
+  res: Response,
+  _next: NextFunction
+) {
   console.error(err);
-  res.status(err.status || 500).send(err.message || 'Something broke!');
-};
+  res.status(err.statusCode || 500).send(err.message || 'Something broke!');
+}